// backend/src/common/enums/rol.enum.ts
export enum RolNombre {
  RESPONSABLE_ALMACEN = 'Responsable de Almacén',
  RESPONSABLE_COMPRAS = 'Responsable de Compras',
  RESPONSABLE_PROYECTOS = 'Responsable de Proyectos',
  RESPONSABLE_TESORERIA = 'Responsable de Tesorería',
  RESPONSABLE_COMERCIALIZACION = 'Responsable de Comercialización y Ventas',
  GERENTE_GENERAL = 'Gerente General',
  ADMINISTRADOR = 'Administrador',
}